+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.x branch
 ## 1.0 branch
 ### 1.0.2
+* added `Tools\TestSuite\TestCaseTrait` with some assertion methods;
+* added `get_class_short_name()` and `get_hostname_from_url()` global functions;
 * fixed `rtr()` global function. It can also use the `ROOT` environment variable.
 
 ### 1.0.1